Put verifyDescription to Detailer page as it is belongs to that. + refactor 2 usages.
